refactor(INJI-435): extract esignetMosipVcItem and existingMosipVcItem machines common functions into separate file
