+++ components/VC/MosipVCItem/VcItemController.tsx
@@ -1,23 +1,20 @@
 import {useContext, useRef} from 'react';
 import {GlobalContext} from '../../../shared/GlobalContext';
 import {
-  createExistingMosipVCItemMachine,
-  ExistingMosipVCItemEvents,
   selectContext,
   selectEmptyWalletBindingId,
   selectGeneratedOn,
-  selectIsSavingFailedInIdle,
   selectKebabPopUp,
   selectVerifiableCredential,
+} from '../../../machines/VCItemMachine/commonSelectors';
+import {
+  createExistingMosipVCItemMachine,
+  ExistingMosipVCItemEvents,
+  selectIsSavingFailedInIdle,
 } from '../../../machines/VCItemMachine/ExistingMosipVCItem/ExistingMosipVCItemMachine';
 import {
   createEsignetMosipVCItemMachine,
   EsignetMosipVCItemEvents,
-  selectContext as esignetSelectContext,
-  selectEmptyWalletBindingId as esignetSelectEmptyWalletBindingId,
-  selectGeneratedOn as esignetSelectGeneratedOn,
-  selectKebabPopUp as esignetSelectKebabPopUp,
-  selectVerifiableCredentials as esignetSelectVerifiableCredentials,
 } from '../../../machines/VCItemMachine/EsignetMosipVCItem/EsignetMosipVCItemMachine';
 import {useInterpret, useSelector} from '@xstate/react';
 import {EsignetMosipVCItemProps, ExistingMosipVCItemProps} from './MosipVCItem';
@@ -50,19 +47,13 @@ export function useVcItemController(
   const storeErrorTranslationPath = 'errors.savingFailed';
   let generatedOn = useSelector(service, selectGeneratedOn);
   if (props.vcMetadata.isFromOpenId4VCI()) {
-    context = useSelector(service, esignetSelectContext);
-    isKebabPopUp = useSelector(service, esignetSelectKebabPopUp);
-    generatedOn = useSelector(service, esignetSelectGeneratedOn);
-    emptyWalletBindingId = useSelector(
-      service,
-      esignetSelectEmptyWalletBindingId,
-    );
+    context = useSelector(service, selectContext);
+    isKebabPopUp = useSelector(service, selectKebabPopUp);
+    generatedOn = useSelector(service, selectGeneratedOn);
+    emptyWalletBindingId = useSelector(service, selectEmptyWalletBindingId);
     DISMISS = () => service.send(EsignetMosipVCItemEvents.DISMISS());
     KEBAB_POPUP = () => service.send(EsignetMosipVCItemEvents.KEBAB_POPUP());
-    verifiableCredential = useSelector(
-      service,
-      esignetSelectVerifiableCredentials,
-    );
+    verifiableCredential = useSelector(service, selectVerifiableCredential);
   }
   return {
     service,
