+++ machines/VCItemMachine/ExistingMosipVCItem/ExistingMosipVCItemMachine.ts
@@ -48,7 +48,7 @@ const model = createModel(
     idType: '' as VcIdType,
     vcMetadata: {} as VCMetadata,
     myVcs: [] as string[],
-    generatedOn: null as Date,
+    generatedOn: new Date() as Date,
     credential: null as DecodedCredential,
     verifiableCredential: null as VerifiableCredential,
     storeVerifiableCredential: null as VerifiableCredential,
@@ -1536,10 +1536,6 @@ export function selectVc(state: State) {
   return data;
 }
 
-export function selectGeneratedOn(state: State) {
-  return new Date(state.context.generatedOn);
-}
-
 export function selectId(state: State) {
   return state.context.vcMetadata.id;
 }
@@ -1552,23 +1548,8 @@ export function selectCredential(state: State) {
   return state.context.credential;
 }
 
-export function selectVerifiableCredential(state: State) {
-  return state.context.verifiableCredential;
-}
-
-export function selectContext(state: State) {
-  return state.context;
-}
-
 export function selectIsOtpError(state: State) {
   return state.context.otpError;
-}
-
-export function selectOtpError(state: State) {
-  return state.context.otpError;
-}
-export function selectIsPinned(state: State) {
-  return state.context.vcMetadata.isPinned;
 }
 
 export function selectIsLockingVc(state: State) {
@@ -1591,38 +1572,12 @@ export function selectIsAcceptingRevokeInput(state: State) {
   return state.matches('acceptingRevokeInput');
 }
 
-export function selectEmptyWalletBindingId(state: State) {
-  var val = state.context.walletBindingResponse
-    ? state.context.walletBindingResponse.walletBindingId
-    : undefined;
-  return val == undefined || val == null || val.length <= 0 ? true : false;
-}
-
-export function selectWalletBindingError(state: State) {
-  return state.context.walletBindingError;
-}
-
-export function selectBindingAuthFailedError(state: State) {
-  return state.context.bindingAuthFailedMessage;
-}
-
 export function selectRequestBindingOtp(state: State) {
   return state.matches('requestingBindingOtp');
 }
 
 export function selectAcceptingBindingOtp(state: State) {
   return state.matches('acceptingBindingOtp');
-}
-
-export function selectShowWalletBindingError(state: State) {
-  return (
-    state.matches('showingWalletBindingError') ||
-    state.matches('kebabPopUp.showingWalletBindingError')
-  );
-}
-
-export function selectWalletBindingSuccess(state: State) {
-  return state.context.walletBindingSuccess;
 }
 
 export function selectWalletBindingInProgress(state: State) {
@@ -1637,41 +1592,9 @@ export function selectWalletBindingInProgress(state: State) {
 export function selectBindingWarning(state: State) {
   return state.matches('showBindingWarning');
 }
-export function selectKebabPopUp(state: State) {
-  return state.matches('kebabPopUp');
-}
 
 export function selectKebabPopUpRequestBindingOtp(state: State) {
   return state.matches('kebabPopUp.requestingBindingOtp');
-}
-
-export function selectKebabPopUpAcceptingBindingOtp(state: State) {
-  return state.matches('kebabPopUp.acceptingBindingOtp');
-}
-
-export function selectKebabPopUpShowWalletBindingError(state: State) {
-  return state.matches('kebabPopUp.showingWalletBindingError');
-}
-
-export function selectKebabPopUpWalletBindingInProgress(state: State) {
-  return state.matches('kebabPopUp.requestingBindingOtp') ||
-    state.matches('kebabPopUp.addingWalletBindingId') ||
-    state.matches('kebabPopUp.addKeyPair') ||
-    state.matches('kebabPopUp.updatingPrivateKey')
-    ? true
-    : false;
-}
-
-export function selectKebabPopUpBindingWarning(state: State) {
-  return state.matches('kebabPopUp.showBindingWarning');
-}
-
-export function selectRemoveWalletWarning(state: State) {
-  return state.matches('kebabPopUp.removeWallet');
-}
-
-export function selectShowActivities(state: State) {
-  return state.matches('kebabPopUp.showActivities');
 }
 
 export function selectIsSavingFailedInIdle(state: State) {
