+++ machines/vc.typegen.ts
@@ -16,19 +16,32 @@ export interface Typegen0 {
     addVcToInProgressDownloads: 'ADD_VC_TO_IN_PROGRESS_DOWNLOADS';
     getReceivedVcsResponse: 'GET_RECEIVED_VCS';
     getVcItemResponse: 'GET_VC_ITEM';
-    loadMyVcs: 'REFRESH_MY_VCS' | 'xstate.init';
+    loadMyVcs:
+      | 'REFRESH_MY_VCS'
+      | 'REMOVE_TAMPERED_VCS'
+      | 'STORE_RESPONSE'
+      | 'xstate.init';
     loadReceivedVcs: 'REFRESH_RECEIVED_VCS' | 'STORE_RESPONSE';
+    logTamperedVCsremoved: 'REMOVE_TAMPERED_VCS';
     moveExistingVcToTop: 'VC_RECEIVED';
     prependToMyVcs: 'VC_ADDED';
     prependToReceivedVcs: 'VC_RECEIVED';
-    removeVcFromInProgressDownlods: 'REMOVE_VC_FROM_IN_PROGRESS_DOWNLOADS';
+    removeDownloadFailedVcsFromStorage: 'DELETE_VC';
+    removeDownloadingFailedVcsFromMyVcs: 'STORE_RESPONSE';
+    removeTamperedVcs: 'REMOVE_TAMPERED_VCS';
+    removeVcFromInProgressDownlods:
+      | 'DOWNLOAD_LIMIT_EXPIRED'
+      | 'REMOVE_VC_FROM_IN_PROGRESS_DOWNLOADS';
     removeVcFromMyVcs: 'REMOVE_VC_FROM_CONTEXT';
     resetAreAllVcsDownloaded: 'RESET_ARE_ALL_VCS_DOWNLOADED';
+    resetDownloadFailedVcs: 'STORE_RESPONSE';
     resetWalletBindingSuccess: 'RESET_WALLET_BINDING_SUCCESS';
     setDownloadedVCFromOpenId4VCI: 'VC_DOWNLOADED_FROM_OPENID4VCI';
     setDownloadedVc: 'VC_DOWNLOADED';
+    setDownloadingFailedVcs: 'DOWNLOAD_LIMIT_EXPIRED';
     setMyVcs: 'STORE_RESPONSE';
     setReceivedVcs: 'STORE_RESPONSE';
+    setTamperedVcs: 'TAMPERED_VC';
     setUpdatedVcMetadatas: 'VC_METADATA_UPDATED';
     setVcUpdate: 'VC_UPDATE';
     setWalletBindingSuccess: 'WALLET_BINDING_SUCCESS';
@@ -40,6 +53,8 @@ export interface Typegen0 {
   };
   eventsCausingServices: {};
   matchesStates:
+    | 'deletingFailedVcs'
+    | 'downloadLimitExpired'
     | 'init'
     | 'init.myVcs'
     | 'init.receivedVcs'
@@ -50,6 +65,7 @@ export interface Typegen0 {
     | 'ready.receivedVcs'
     | 'ready.receivedVcs.idle'
     | 'ready.receivedVcs.refreshing'
+    | 'tamperedVCs'
     | {
         init?: 'myVcs' | 'receivedVcs';
         ready?:
